pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opc-cli"
version = "0.1.0"
description = "OPC 工具集：B站视频转写 + 语音合成 CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "zhipuai>=2.0.0",
    "yt-dlp>=2024.0.0",
    "soundfile>=0.12.0",
    "numpy>=1.20.0",
]

[project.scripts]
opc = "opc_cli.cli:main"

[tool.setuptools.packages.find]
include = ["opc_cli*"]
